% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest-dev-obs.R
\name{latest_device_observation}
\alias{latest_device_observation}
\title{Get latest device observation}
\usage{
latest_device_observation(device_id, pat = tempest_pat())
}
\arguments{
\item{device_id}{(chr) device id}

\item{pat}{see \code{\link[=tempest_pat]{tempest_pat()}}}
}
\value{
list with device metadata, configured, units and point-in-time observation
}
\description{
Get the latest observation from one of your devices.
}
\details{
The observations are not named but are in this order:
}
\examples{
\dontrun{
st <- tempest_station_meta()
latest_device_observation(st$stations$devices[[1]]$device_id[2])$obs
}
}
\references{
\href{https://weatherflow.github.io/Tempest/api/}{Tempest API}
}
