% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{latest_forecast}
\alias{latest_forecast}
\title{Get latest forecast
#'}
\usage{
latest_forecast(station_id, pat = tempest_pat())
}
\arguments{
\item{station_id}{(chr) station id}

\item{pat}{see \code{\link[=tempest_pat]{tempest_pat()}}}
}
\value{
list with station metadata, configured, units and forecast information
}
\description{
Get latest forecast
#'
}
\examples{
\dontrun{
st <- tempest_station_meta()
latest_observation(st$stations$station_id)
}
}
\references{
\href{https://weatherflow.github.io/Tempest/api/}{Tempest API}
}
