% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest-obs.R
\name{latest_observation}
\alias{latest_observation}
\title{Get latest station observation}
\usage{
latest_observation(station_id, pat = tempest_pat())
}
\arguments{
\item{station_id}{(chr) station id}

\item{pat}{see \code{\link[=tempest_pat]{tempest_pat()}}}
}
\value{
list with station metadata, configured, units and point-in-time observation
}
\description{
Get the latest most recent observation for a station.
}
\examples{
\dontrun{
st <- tempest_station_meta()
latest_observation(st$stations$station_id)
}
}
\references{
\href{https://weatherflow.github.io/Tempest/api/}{Tempest API}
}
