% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{tempest_pat}
\alias{tempest_pat}
\title{Get or set TEMPEST_PAT value}
\usage{
tempest_pat(force = FALSE)
}
\arguments{
\item{force}{force setting a new Tempest PAT for the current environment?}
}
\value{
atomic character vector containing the Tempest Weather PAT
}
\description{
The API wrapper functions in this package all rely on a Tempest Weather
personal access token residing in the environment variable \code{TEMPEST_PAT}. The
easiest way to accomplish this is to set it in the \verb{\code{.Renviron}} file in your
home directory.
}
