% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{tempest_station_meta}
\alias{tempest_station_meta}
\title{Get station metadata}
\usage{
tempest_station_meta(pat = tempest_pat())
}
\arguments{
\item{pat}{see \code{\link[=tempest_pat]{tempest_pat()}}}
}
\value{
list with station metadata
}
\description{
Retrieve a list of your stations along with all connected devices.
}
\examples{
# tempest_station_meta()
}
\references{
\url{https://weatherflow.github.io/Tempest/api/}
}
