% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempestwx-package.R
\docType{package}
\name{tempestwx}
\alias{tempestwx}
\alias{tempestwx-package}
\title{Tools To Query Tempest Weather Station Data}
\description{
The Tempest Weather System combines state-of-the-art sensors and a sleek,
wireless design with WeatherFlow’s proprietary modeling capabilities to seamlessly
present validated weather data and improved forecasts. The Tempest API & Developer
Platform empowers a growing community of developers to build useful applications and
integrations by providing access to enhanced data and forecasts from the Tempest System.
Tools are provided to query Tempest Weather station data.
}
\references{
\url{https://weatherflow.github.io/Tempest/api/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/tempestwx}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/tempestwx/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
