% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{udp_logger}
\alias{udp_logger}
\title{Log local UDP Tempest messages to a file}
\usage{
udp_logger(path)
}
\arguments{
\item{path}{full path (no expansion is performed) to a file}
}
\value{
nothing
}
\description{
This dedicates the R session to logging UDP packets to a file. No other operations
can be performed until terminated.
}
\examples{
# udp_logger("/tmp/logger.json")
}
\references{
\url{https://weatherflow.github.io/Tempest/api/}
}
