% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempestwx-package.R
\docType{package}
\name{tempestwx}
\alias{tempestwx}
\alias{tempestwx-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/tempestwx}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/tempestwx/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
